Dedupe styles
- Remove `defaultTheme` from `@splitgraph/design` in favor of `themes/index.ts`
   from `@splitgraph/tdesign`
- Use the debugging tool from the last commit to figure out which rules were only
   in `@splitgraph/design`, and move them to `@splitgraph/tdesign`
- History: We started with `@splitgraph/design`, but then switched to TS and started
  using `@splitgraph/tdesign`. But there were also hot reload / code sharing issues
  that made things confusing, and e.g. the docs pages still only used `@splitgraph/design`
- In many places we were using two nested `ThemeProvider`s, one for each theme
- For the most part, the dedupe was easy since most new style rules were added to
  `@splitgraph/tdesign`. Only a few values had to change
- I've checked most pages of the site and I think we're okay. And the debugging tool
   says there area zero style rules that only `@splitgraph/design` has. The maths are there

(still some WIP/cleanup to do, but got rid of the main problem)

(this is step 1 of the redesign, now at least we can take an accurate inventory of what we have)
