+++ docs/hocs/withTheme.js
@@ -1,18 +1,20 @@
 import { getDisplayName } from "next/dist/next-server/lib/utils";
 
 // This is kinda hacky atm, and should be a more unified export from tcomponents
-import { ThemeProvider as OurThemeProvider } from "theme-ui";
+// import { ThemeProvider as OurThemeProvider } from "theme-ui";
 import { ThemeProvider, makeDefaultTheme } from "@splitgraph/tdesign";
 
-const defaultTheme = !!makeDefaultTheme ? makeDefaultTheme() : {};
+// const defaultTheme = !!makeDefaultTheme ? makeDefaultTheme() : {};
+
+const defaultTheme = makeDefaultTheme();
 
 const withTheme = (Page) => {
   const WithTheme = (props) => (
-    <OurThemeProvider theme={defaultTheme}>
-      <ThemeProvider theme={defaultTheme}>
-        <Page {...props} />
-      </ThemeProvider>
-    </OurThemeProvider>
+    // <OurThemeProvider theme={defaultTheme}>
+    <ThemeProvider theme={defaultTheme}>
+      <Page {...props} />
+    </ThemeProvider>
+    // </OurThemeProvider>
   );
 
   if (Page.getInitialProps) {
